projects/rag-chatbot/requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
